requirements.txt
opencv-python==4.7.0.72
pycryptodomex==3.14.1
